refactor: optimize Offer page and restructure components by FSD
