+++ src/app/store/rootReducer.ts
@@ -3,8 +3,8 @@ import { offerListReducer } from '../../entities/offer/model/offersListSlice';
 import { userReducer } from '../../entities/user/model/userSlice';
 import { citySelectorReducer } from '../../features/CitySelector/model/citySelectorSlice';
 import { favoritesReducer } from '../../features/Favorites/model/favoritesSlice';
-import { offerDetailsReducer } from '../../features/OfferContent/model/offerDetailsSlice';
 import { sortOffersReducer } from '../../features/SortOffers/model/sortOffersSlice';
+import { offerDetailsReducer } from '../../widgets/OfferContent/model/offerDetailsSlice';
 
 export const rootReducer = combineReducers({
   user: userReducer,
