+++ src/widgets/OfferContent/index.ts
@@ -1,1 +1,2 @@
+export { OfferContent } from './ui/OfferContent';
 export { OfferDetails } from './ui/OfferDetails';
