+++ src/widgets/OfferContent/model/offer-details.thunks.ts
